fix: move resource query parsing to overlay entry to ensure its success
