+++ client/utils/errorEventHandlers.js
