+++ client/utils/formatWebpackErrors.js
